Renamed common.py to util.py
